Divided the main component in two
